+++ rbx_reflection_database/CHANGELOG.md
@@ -3,9 +3,11 @@
 ## Unreleased
 * Aliased `StarterPlayer.AvatarJointUpgrade` to `StarterPlayer.AvatarJointUpgrade_SerializedRollout`. ([#599])
 * Updated to Roblox version 718. ([#604])
+* Added migrations for `UICorner.CornerRadius`. ([#612])
 
 [#599]: https://github.com/rojo-rbx/rbx-dom/pull/599
 [#604]: https://github.com/rojo-rbx/rbx-dom/pull/604
+[#612]: https://github.com/rojo-rbx/rbx-dom/pull/612
 
 ## 2.0.2+roblox-700 (2025-11-27)
 * Updated to Roblox version 700.